CHORE: 자잘한 디테일 변경 및 버그 해결
- 부정보의 등급을 주정보로 옮기기
- 전체 강화, 개인 강화에 따른 추가값을 따로 분류하기

+++ Data/BuffData.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 public class BuffData
 {
     public int      id;             // 버프 id
-    public int      value;          // 값
+    public float    value;          // 값
     public string   descripition;   // 설명
 }
